+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
 * json definitions added
 * validate_verbose function
+* schema validation improvements
 
 
 0.1 (2017-08-01)